feat: 新增elasticsearch搜索方案，新增每日全量同步数据库和es的定时任务

+++ src/main/java/io/github/novel/mynovel/service/impl/BookServiceImpl.java
@@ -410,7 +410,7 @@ public class BookServiceImpl implements BookService {
 
         // 构建小说实体
         BookInfo bookInfo = new BookInfo();
-        bookInfo.setWorkDirection(dto.getWorkDirection().byteValue());
+        bookInfo.setWorkDirection(dto.getWorkDirection());
         bookInfo.setCategoryId(dto.getCategoryId());
         bookInfo.setCategoryName(dto.getCategoryName());
         bookInfo.setPicUrl(dto.getPicUrl());
